WIP Impling on demand sync - build failing b/c of no provider for ChatMessagesDao.kt

+++ app/src/main/java/com/nextcloud/talk/data/changeListVersion/SyncableModel.kt
@@ -16,7 +16,7 @@ interface SyncableModel {
     /**
      * Model identifier.
      */
-    var changedId: Int
+    var changedId: Long
 
     /**
      * Model deletion checker.
