+++ app/src/main/java/com/nextcloud/talk/models/json/chat/ChatMessage.kt
@@ -146,7 +146,7 @@ data class ChatMessage(
 
     var openWhenDownloaded: Boolean = true,
 
-    override var changedId: Int = jsonMessageId,
+    override var changedId: Long = jsonMessageId.toLong(),
 
     // override var markedForDeletion: Boolean = "comment_deleted" == messageType
     override var markedForDeletion: Boolean = false
